backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
certifi==2024.8.30
cffi==2.0.0
click==8.1.7
colorama==0.4.6
cryptography==46.0.2
deprecation==2.1.0
fastapi==0.118.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==24.2
postgrest==2.21.1
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
realtime==2.21.1
sniffio==1.3.1
starlette==0.45.2
storage3==2.21.1
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
